+++ src/tamtam/controller.py
@@ -1,18 +1,19 @@
 import asyncio
+from tamtam.socket import TTMobileServer
 from classes.controllerbase import ControllerBase
 from common.config import ServerConfig
-from tamtam_ws.server import TTWSServer
 
-class TTWSController(ControllerBase):
+class TTMobileController(ControllerBase):
     def __init__(self):
         self.config = ServerConfig()
 
     def launch(self, api):
         async def _start_all():
             await asyncio.gather(
-                TTWSServer(
+                TTMobileServer(
                     host=self.config.host,
-                    port=self.config.tamtam_ws_port,
+                    port=self.config.tamtam_tcp_port,
+                    ssl_context=api['ssl'],
                     db_pool=api['db'],
                     clients=api['clients'],
                     send_event=api['event']
