+++ src/tamtam/socket.py
@@ -1,7 +1,8 @@
 import asyncio, logging, traceback
-from tamtam_tcp.proto import Proto
-from tamtam_tcp.processors import Processors
+from common.proto_tcp import MobileProto
+from tamtam.processors import Processors
 from common.rate_limiter import RateLimiter
+from common.opcodes import Opcodes
 
 class TTMobileServer:
     def __init__(self, host="0.0.0.0", port=443, ssl_context=None, db_pool=None, clients={}, send_event=None):
@@ -13,7 +14,9 @@ class TTMobileServer:
         self.db_pool = db_pool
         self.clients = clients
 
-        self.proto = Proto()
+        self.opcodes = Opcodes()
+
+        self.proto = MobileProto()
         self.processors = Processors(db_pool=db_pool, clients=clients, send_event=send_event)
 
         # rate limiter
@@ -69,28 +72,28 @@ class TTMobileServer:
                 payload = packet.get("payload")
 
                 match opcode:
-                    case self.proto.HELLO:
-                        deviceType, deviceName = await self.processors.process_hello(payload, seq, writer)
-                    case self.proto.REQUEST_CODE:
+                    case self.opcodes.SESSION_INIT:
+                        deviceType, deviceName = await self.processors.session_init(payload, seq, writer)
+                    case self.opcodes.AUTH_REQUEST:
                         if not self.auth_rate_limiter.is_allowed(address[0]):
-                            await self.processors._send_error(seq, self.proto.REQUEST_CODE, self.processors.error_types.RATE_LIMITED, writer)
+                            await self.processors._send_error(seq, self.opcodes.AUTH_REQUEST, self.processors.error_types.RATE_LIMITED, writer)
                         else:
-                            await self.processors.process_request_code(payload, seq, writer)
-                    case self.proto.VERIFY_CODE:
+                            await self.processors.auth_request(payload, seq, writer)
+                    case self.opcodes.AUTH:
                         if not self.auth_rate_limiter.is_allowed(address[0]):
-                            await self.processors._send_error(seq, self.proto.VERIFY_CODE, self.processors.error_types.RATE_LIMITED, writer)
+                            await self.processors._send_error(seq, self.opcodes.AUTH, self.processors.error_types.RATE_LIMITED, writer)
                         else:
-                            await self.processors.process_verify_code(payload, seq, writer)
-                    case self.proto.FINAL_AUTH:
+                            await self.processors.auth(payload, seq, writer)
+                    case self.opcodes.AUTH_CONFIRM:
                         if not self.auth_rate_limiter.is_allowed(address[0]):
-                            await self.processors._send_error(seq, self.proto.FINAL_AUTH, self.processors.error_types.RATE_LIMITED, writer)
+                            await self.processors._send_error(seq, self.opcodes.AUTH_CONFIRM, self.processors.error_types.RATE_LIMITED, writer)
                         else:
-                            await self.processors.process_final_auth(payload, seq, writer, deviceType, deviceName)
-                    case self.proto.LOGIN:
+                            await self.processors.auth_confirm(payload, seq, writer, deviceType, deviceName)
+                    case self.opcodes.LOGIN:
                         if not self.auth_rate_limiter.is_allowed(address[0]):
-                            await self.processors._send_error(seq, self.proto.LOGIN, self.processors.error_types.RATE_LIMITED, writer)
+                            await self.processors._send_error(seq, self.opcodes.LOGIN, self.processors.error_types.RATE_LIMITED, writer)
                         else:
-                            userPhone, userId, hashedToken = await self.processors.process_login(payload, seq, writer)
+                            userPhone, userId, hashedToken = await self.processors.login(payload, seq, writer)
 
                             if userPhone:
                                 await self._finish_auth(writer, address, userPhone, userId)
